+++ CHANGELOG.md
@@ -8,6 +8,27 @@
 *Tiledesk SRL*
 
 
+# 1.39.9-rc2:
+- **added**: set default model
+- **changed**: replaced the textarea with a select in the model selection
+- **changed**: changed token display by adding k for thousands
+- **changed**: editable max token field
+- **added**: added max_token and min_token parameters for each model
+
+
+# 1.39.9-rc1:
+- **added**: added vLLM model integration
+
+
+# 1.39.8-rc1:
+- **changed**: set max_tokens from 8192 to 100000 only for gpt-5 model
+- **added**: added css for temperature disabled when gpt-5 is the selected model 
+- **added**: set max_token when change model
+
+
+# 1.39.6
+- **changed**: set max_tokens from 8192 to 100000
+
 # 1.39.6
 - **changed**: set max_tokens from 8192 to 100000